sum of non-prime index values in an array
A prime number (or a prime) is a natural number greater than 1 that has no positive divisors other than 1 and itself. In other words, a prime number is a whole number greater than 1, whose only two whole-number factor are 1 and itself. The first few prime number are 2,3,5,7,11,13,17,19,23 and 29.
Given an array with ‘N’ elements, you are expected to find the sum of the values that are present in non-prime indexes of the array. Note that the array index starts with 0 i.e, the position (index) of the first array element is 0, the position of the next array element is 1, and so on.